visualizer/pyproject.toml
[tool.poetry]
name = "cleaning-visualizer"
version = "0.1.0"
description = "掃除管理システムの可視化ダッシュボード"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "cleaning_visualizer"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.28.0"
gspread = "^5.12.0"
google-auth = "^2.23.4"
pandas = "^2.1.3"
plotly = "^5.17.0"
numpy = "^1.24.3"
python-dateutil = "^2.8.2"
pytz = "^2023.3"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cleaning-visualizer = "cleaning_visualizer.app:main" 